refactor: mid work
